refactor directory in util
